Conexiones terminadas

+++ database/flask_login.sql
@@ -46,7 +46,7 @@ CREATE TABLE envios (
   `nombre` varchar(50) COLLATE utf8_unicode_ci NOT NULL,
   `email` varchar(50) COLLATE utf8_unicode_ci NOT NULL,
   `telefono` varchar(15) COLLATE utf8_unicode_ci NOT NULL,
-  `costo` INTEGER COLLATE utf8_unicode_ci NOT NULL,
+  `costo` FLOAT COLLATE utf8_unicode_ci NOT NULL,
   `idusuario` INTEGER COLLATE utf8_unicode_ci NOT NULL,
   `idrepartidor` INTEGER COLLATE utf8_unicode_ci NOT NULL,
   PRIMARY KEY(`id`)
